fix: decode float4 as the number the server would print
A float4 arrives as four bytes and widens into a double that spells its
binary approximation out in full, so the column PostgreSQL prints as 1.1
read back as 1.100000023841858. The text path never had the problem - the
server shortens before sending - so the same column answered differently
depending on a wire format the caller may not know is there, and binary,
our own default for the extended query path, was the wrong-looking one:

    1.1::float4      binary 1.100000023841858     text 1.1
    0.1::float4      binary 0.10000000149011612   text 0.1
    3.14159::float4  binary 3.141590118408203     text 3.14159

The binary decode now finds the shortest decimal that reads back as the
same float4. Nine significant digits always round-trip a binary32 and
most need far fewer, so it rounds to successively longer forms and takes
the first that reads back unchanged. Integers within +-2^24 are exact in
float4 and skip the search, but only within it - 1e20 widens to
100000002004087730000 and has to be shortened like anything else.

Rounding by hand, rather than through toPrecision(), is what makes it
affordable: 24ns a value against 115ns, because toPrecision() costs four
times what the engine's own shortest-form toString does and no built-in
is cheaper. Scaling by a power of ten is only faithful while both the
digits and the power are exact in a double, so a magnitude check keeps
the power inside its own exact range and a value sitting exactly halfway
goes to the string path as well - Math.round() breaks those ties away
from zero where dtoa breaks them to even.

Checked against the string path over 502,150 values, every exponent
crossed with a spread of mantissas then half a million random bit
patterns: the same answer every time. Negative zero is worth a mention,
since the tests pin it - toPrecision() drops the sign, so the string
route alone would have turned -0 into +0.

Four existing tests pinned the old values through Math.fround(); they
expect the shortened form now. The two float4 parse tests either side of
them used 1.25 and -2.5, both exact in float4, which is why none of this
showed.

+++ test/B-connection/04-query.spec.ts
@@ -276,9 +276,11 @@ describe('query() (Extended Query)', () => {
     expect(row.f_int2).toStrictEqual(1);
     expect(row.f_int4).toStrictEqual(12345);
     expect(row.f_int8).toStrictEqual(BigInt('9007199254740995'));
-    // float4 is IEEE754 single-precision; 1.2 is not exactly representable
-    // in 32 bits, so the stored value widens to the nearest float4 value.
-    expect(row.f_float4).toStrictEqual(Math.fround(1.2));
+    // 1.2 is not exactly representable in 32 bits, but it is the shortest
+    // decimal that reads back as the float4 that was stored, so that is
+    // what comes out - the number the server prints, not the double that
+    // spells its binary approximation out in full.
+    expect(row.f_float4).toStrictEqual(1.2);
     expect(row.f_float8).toStrictEqual(5.12345);
     expect(row.f_char).toStrictEqual('a');
     expect(row.f_varchar).toStrictEqual('abcd');
